requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cshogi==0.9.7
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fonttools==4.61.1
git-filter-repo==2.47.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
plotdigitizer==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-shogi==1.1.1
pywin32-ctypes==0.2.3
reportlab==4.4.10
requests==2.32.5
rsa==4.9.1
selenium==4.39.0
setuptools==80.10.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
tenacity==9.1.2
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wheel==0.46.3
wsproto==1.3.2
